+++ application/SaLsA/de.uniol.inf.is.odysseus.classification/src/de/uniol/inf/is/odysseus/classification/segmentation/SEFSegmentation.java
@@ -25,7 +25,7 @@ public class SEFSegmentation extends AbstractFunction<Geometry> {
     private static final long           serialVersionUID = 4973220087210280849L;
 
     public static final SDFDatatype[][] accTypes         = new SDFDatatype[][] {
-            { SDFSpatialDatatype.SPATIAL_MULTI_POINT }, { SDFDatatype.DOUBLE }, { SDFDatatype.DOUBLE } };
+            { SDFSpatialDatatype.SPATIAL_MULTI_POINT }, { SDFDatatype.DOUBLE }, { SDFDatatype.INTEGER } };
     private final GeometryFactory       geometryFactory  = new GeometryFactory();
 
     @Override
